Extract tests for isSet to own file
