+++ lib/referee.test.js
@@ -4,37 +4,6 @@ var referee = require("../lib/referee");
 var testHelper = require("./test-helper");
 var sinon = require("sinon");
 
-testHelper.assertionTests("assert", "isSet", function(pass, fail, msg, error) {
-    function captureArgs() {
-        return arguments;
-    }
-
-    pass("for Set", new Set());
-    fail("for string", "apple pie");
-    fail("for array", []);
-    fail("for object", {});
-    fail("for arguments", captureArgs());
-    msg(
-        "fail with descriptive message",
-        "[assert.isSet] Expected {  } to be a Set",
-        {}
-    );
-    msg(
-        "fail with custom message",
-        "[assert.isSet] Nope: Expected {  } to be a Set",
-        {},
-        "Nope"
-    );
-    error(
-        "for object",
-        {
-            code: "ERR_ASSERTION",
-            operator: "assert.isSet"
-        },
-        {}
-    );
-});
-
 testHelper.assertionTests("assert", "isSymbol", function(
     pass,
     fail,
